Tests/LibWeb/Text/expected/css/root-color-scheme-loading-canvas.txt

Ready state: loading
Computed root color-scheme: light
Canvas color-scheme: light
